nn_dl/src/requirements.txt
flask
python-dotenv
pyyaml
matplotlib
numpy
tensorflow
scikit-image
tensorflow.keras
